factor out str segment to util
