+++ komi_syntax/src/ast/mod.rs
@@ -1,4 +1,5 @@
 use komi_util::Range;
+pub use komi_util::{StrSegment, StrSegmentKind};
 
 /// Kinds of AST produced during parsing.
 /// Serves as the interface between a parser and its user.
@@ -7,6 +8,7 @@ pub enum AstKind {
     Program { expressions: Vec<Box<Ast>> },
     Number(f64),
     Bool(bool),
+    Str(Vec<StrSegment>),
     Identifier(String),
     PrefixPlus { operand: Box<Ast> },
     PrefixMinus { operand: Box<Ast> },
