+++ komi_syntax/src/token/mod.rs
@@ -1,4 +1,5 @@
 use komi_util::Range;
+pub use komi_util::{StrSegment, StrSegmentKind};
 
 /// A token produced during lexing.
 #[derive(Debug, PartialEq, Clone)]
@@ -85,38 +86,6 @@ pub enum TokenKind {
     ElseBranch,
     /// An iteration keyword `반복`.
     Iteration,
-}
-
-/// A string segment in a string token.
-#[derive(Debug, PartialEq, Clone)]
-pub struct StrSegment {
-    pub kind: StrSegmentKind,
-    pub location: Range,
-}
-
-impl StrSegment {
-    pub fn new(kind: StrSegmentKind, location: Range) -> Self {
-        Self { kind, location }
-    }
-}
-
-/// A kind of string segment in a string token.
-#[derive(Debug, PartialEq, Clone)]
-pub enum StrSegmentKind {
-    /// A string segment, such as `사과` in "`사과{오렌지}`".
-    Str(String),
-    /// An interpolated identifier, such as `오렌지` in "`사과{오렌지}`".
-    Identifier(String),
-}
-
-impl StrSegmentKind {
-    pub fn str(s: impl Into<String>) -> Self {
-        Self::Str(s.into())
-    }
-
-    pub fn identifier(s: impl Into<String>) -> Self {
-        Self::Identifier(s.into())
-    }
 }
 
 /// Makes a token with the kind and the location.
