+++ komi_util/src/lib.rs
@@ -3,6 +3,7 @@ pub mod environment;
 pub mod error;
 pub mod location;
 pub mod scanner;
+pub mod str_segment;
 pub mod tape;
 pub mod unpacker;
 
@@ -12,4 +13,5 @@ pub use location::range;
 pub use location::range::Range;
 pub use location::spot::Spot;
 pub use scanner::Scanner;
+pub use str_segment::{StrSegment, StrSegmentKind};
 pub use tape::Tape;
