+++ komi_util/src/str_segment/mod.rs
@@ -1,0 +1,36 @@
+//! StrSegment
+//! A segment of a string.
+
+use crate::Range;
+
+/// A string segment in a string token.
+#[derive(Debug, PartialEq, Clone)]
+pub struct StrSegment {
+    pub kind: StrSegmentKind,
+    pub location: Range,
+}
+
+impl StrSegment {
+    pub fn new(kind: StrSegmentKind, location: Range) -> Self {
+        Self { kind, location }
+    }
+}
+
+/// A kind of string segment in a string token.
+#[derive(Debug, PartialEq, Clone)]
+pub enum StrSegmentKind {
+    /// A string segment, such as `사과` in "`사과{오렌지}`".
+    Str(String),
+    /// An interpolated identifier, such as `오렌지` in "`사과{오렌지}`".
+    Identifier(String),
+}
+
+impl StrSegmentKind {
+    pub fn str(s: impl Into<String>) -> Self {
+        Self::Str(s.into())
+    }
+
+    pub fn identifier(s: impl Into<String>) -> Self {
+        Self::Identifier(s.into())
+    }
+}
